Incorporate script to limit logs on travis to the logs preceding the build failure
The build exceeds the 4MB limit for Travis.  When there is a failure, there is no way to troubleshoot since the logs are cut off.  The script records the log and then tails it when there is an error so the 4MB limit won't be exceeded.  http://stackoverflow.com/questions/26082444/how-to-work-around-travis-cis-4mb-output-limit